requirements.txt
streamlit
openai
llama-index
nltk
pypdf
python-dotenv
pdfminer.six